requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai>=1.10.0,<2.0.0
chromadb==0.4.18
PyGithub==1.59.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
